improve design and modularity
