+++ .env.example
@@ -1,3 +1,9 @@
 DATABASE_URL="postgresql://john-doe"
 NEXTAUTH_SECRET="something"
 MIDDLEWARE_SECRET="something"
+
+GITHUB_CLIENT_ID=""
+GITHUB_CLIENT_SECRET=""
+
+GOOGLE_CLIENT_ID=""
+GOOGLE_CLIENT_SECRET=""
