+++ src/app/login/GithubLoginButton.tsx
@@ -1,0 +1,35 @@
+import { Button } from "@/components/ui/button";
+
+const GithubLoginButton = () => {
+  return (
+    <Button
+      variant="outline"
+      className="flex items-center rounded-3xl text-zinc-900"
+    >
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        xmlnsXlink="http://www.w3.org/1999/xlink"
+        className="mr-2"
+        width="20px"
+        height="40px"
+        viewBox="0 0 20 19"
+        version="1.1"
+      >
+        <g id="surface3">
+          <path
+            style={{
+              stroke: "none",
+              fillRule: "evenodd",
+              fill: "rgb(14.117647%,16.078431%,18.431373%)",
+              fillOpacity: "1",
+            }}
+            d="M 10.074219 0 C 4.503906 0 0 4.398438 0 9.84375 C 0 14.195312 2.886719 17.878906 6.886719 19.179688 C 7.386719 19.28125 7.570312 18.96875 7.570312 18.710938 C 7.570312 18.480469 7.554688 17.699219 7.554688 16.882812 C 4.753906 17.46875 4.167969 15.710938 4.167969 15.710938 C 3.71875 14.570312 3.050781 14.277344 3.050781 14.277344 C 2.132812 13.671875 3.117188 13.671875 3.117188 13.671875 C 4.136719 13.738281 4.667969 14.683594 4.667969 14.683594 C 5.570312 16.183594 7.019531 15.757812 7.605469 15.5 C 7.6875 14.863281 7.953125 14.421875 8.238281 14.179688 C 6.003906 13.949219 3.652344 13.101562 3.652344 9.320312 C 3.652344 8.246094 4.050781 7.367188 4.6875 6.683594 C 4.585938 6.4375 4.234375 5.425781 4.785156 4.074219 C 4.785156 4.074219 5.636719 3.8125 7.554688 5.085938 C 8.375 4.867188 9.222656 4.757812 10.074219 4.757812 C 10.921875 4.757812 11.789062 4.871094 12.589844 5.085938 C 14.507812 3.8125 15.359375 4.074219 15.359375 4.074219 C 15.910156 5.425781 15.558594 6.4375 15.460938 6.683594 C 16.109375 7.367188 16.492188 8.246094 16.492188 9.320312 C 16.492188 13.101562 14.140625 13.933594 11.890625 14.179688 C 12.257812 14.488281 12.574219 15.074219 12.574219 16.003906 C 12.574219 17.324219 12.558594 18.382812 12.558594 18.707031 C 12.558594 18.96875 12.742188 19.28125 13.242188 19.179688 C 17.242188 17.878906 20.128906 14.195312 20.128906 9.84375 C 20.144531 4.398438 15.625 0 10.074219 0 Z M 10.074219 0 "
+          />
+        </g>
+      </svg>
+      Sign in with GitHub
+    </Button>
+  );
+};
+
+export default GithubLoginButton;
